requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
Django==4.2.7
django-cors-headers==4.3.0
django-encrypted-fields==1.1.2
django-encrypted-model-fields==0.6.5
djangorestframework==3.14.0
idna==3.4
Pillow==10.1.0
pyasn1==0.5.0
pycparser==2.21
pycrypto==2.6.1
python-keyczar==0.716
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
whitenoise==6.6.0
